+++ src/examples/wikidata.ts
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom';
 
 import {
     Workspace, WorkspaceProps, SparqlDataProvider, OrganizationTemplate, DefaultElementTemplate, PersonTemplate,
-    WikidataSettings, SparqlQueryMethod,
+    WikidataSettings, SparqlQueryMethod, PropertySuggestionParams, PropertyScore,
 } from '../index';
 
 import { onPageLoad, tryLoadLayoutFromLocalStorage, saveLayoutToLocalStorage } from './common';
@@ -11,8 +11,77 @@ import { onPageLoad, tryLoadLayoutFromLocalStorage, saveLayoutToLocalStorage } f
 require('jointjs/css/layout.css');
 require('jointjs/css/themes/default.css');
 
-function onWorkspaceMounted(workspace: Workspace) {
-    if (!workspace) { return; }
+const WIKIDATA_PREFIX = 'http://www.wikidata.org/prop/direct/';
+
+let workspace: Workspace;
+
+function getElementLabel(id: string): string {
+    const model = workspace.getModel();
+    const diagram = workspace.getDiagram();
+    const element = model.getElement(id);
+    return element ? diagram.getLocalizedText(element.template.label.values).text : '';
+}
+
+function wikidataSuggestProperties(params: PropertySuggestionParams) {
+    const idMap: { [id: string]: string } = {};
+
+    const properties = params.properties.map(id => {
+        let resultID;
+        if (id.startsWith(WIKIDATA_PREFIX)) {
+            resultID = id.substr(WIKIDATA_PREFIX.length, id.length);
+        } else {
+            resultID = id;
+        }
+        idMap[resultID] = id;
+        return resultID;
+    });
+    const term = params.token.toLowerCase() || getElementLabel(params.elementId);
+    const requestBody = {
+        threshold: 0.1,
+        term,
+        instance_properties: properties,
+    };
+    return fetch('/wikidata-prop-suggest', {
+        method: 'POST',
+        body: JSON.stringify(requestBody),
+        credentials: 'same-origin',
+        mode: 'cors',
+        cache: 'default',
+    }).then((response) => {
+        if (response.ok) {
+            return response.json();
+        } else {
+            const error = new Error(response.statusText);
+            (error as any).response = response;
+            throw error;
+        }
+    }).then(json => {
+        const dictionary: { [id: string]: PropertyScore } = {};
+        for (const term of json.data) {
+            const propertyIri = idMap[term.id];
+            const item = dictionary[propertyIri];
+
+            if (item && item.score > term.value) { continue; }
+
+            dictionary[propertyIri] = {propertyIri, score: term.value};
+        }
+
+        Object.keys(idMap).forEach(key => {
+            const propertyIri = idMap[key];
+
+            if (dictionary[propertyIri]) { return; }
+
+            dictionary[propertyIri] = {propertyIri, score: 0};
+        });
+
+        return dictionary;
+    });
+}
+
+function onWorkspaceMounted(wspace: Workspace) {
+    if (!wspace) { return; }
+
+    workspace = wspace;
 
     const diagram = workspace.getDiagram();
     diagram.registerTemplateResolver(types => {
@@ -64,6 +133,9 @@ const props: WorkspaceProps & ClassAttributes<Workspace> = {
         const {layoutData} = workspace.getModel().exportLayout();
         window.location.hash = saveLayoutToLocalStorage(layoutData);
         window.location.reload();
+    },
+    viewOptions: {
+        suggestProperties: wikidataSuggestProperties,
     },
 };